+++ include/mppic/critics/path_angle_critic.hpp
@@ -25,7 +25,7 @@ protected:
   size_t offset_from_furthest_{0};
 
   unsigned int power_{0};
-  double weight_{0};
+  float weight_{0};
 };
 
 }  // namespace mppi::critics
