+++ include/mppic/models/optimizer_settings.hpp
@@ -12,8 +12,8 @@ struct OptimizerSettings
   models::ControlConstraints base_constraints{0, 0, 0, 0};
   models::ControlConstraints constraints{0, 0, 0, 0};
   models::SamplingStd sampling_std{0, 0, 0};
-  double model_dt{0};
-  double temperature{0};
+  float model_dt{0};
+  float temperature{0};
   unsigned int batch_size{0};
   unsigned int time_steps{0};
   unsigned int iteration_count{0};
